Add pod example
A pod is an unit of a cluster. In a pod is possible allocate one or more
containers.

Examples of use:

kubectl apply -f pod.yaml

To see the pod status:

kubectl get pods

To test the pod:

kubectl port-forward pod/goserver 8080:80

In the above example 8080 is the host port. 80 is the container port.